requirements.txt
urllib3>=1.22
Django>=1.8
djangorestframework>=3.0